requirements.txt
telethon
google-generativeai
aiogram
httpx
python-dotenv
tenacity
Pillow
beautifulsoup4
